Changes extensions handling mechanism.
Initially, there was no way to configure the extension (be it a handler,
check, etc.) in the same way as a normal component, by defining it in
the config files. THis change enables that, by searching for an
'extension' attribute in the definition and, if found, passing that
definition to the extension. This allows for more versatile extensions.